requirements.txt
biopython==1.79
pandas==1.3.4
numpy==1.19.5
seaborn==0.11.2
matplotlib==3.5.1
requests==2.27.1
re==2.2.1
multi-key-dict==2.0.3